Factor out ElementResolver
