+++ src/ActionResolver.php
@@ -11,22 +11,20 @@ use webignition\BasilModelProvider\Page\PageProviderInterface;
 use webignition\BasilModels\Action\ActionInterface;
 use webignition\BasilModels\Action\InputActionInterface;
 use webignition\BasilModels\Action\InteractionActionInterface;
-use webignition\BasilModels\ElementReference\ElementReference;
-use webignition\BasilModels\PageElementReference\PageElementReference;
 
 class ActionResolver
 {
-    private $pageElementReferenceResolver;
+    private $elementResolver;
 
-    public function __construct(PageElementReferenceResolver $pageElementReferenceResolver)
+    public function __construct(ElementResolver $referencedElementResolver)
     {
-        $this->pageElementReferenceResolver = $pageElementReferenceResolver;
+        $this->elementResolver = $referencedElementResolver;
     }
 
     public static function createResolver(): ActionResolver
     {
         return new ActionResolver(
-            PageElementReferenceResolver::createResolver()
+            ElementResolver::createResolver()
         );
     }
 
@@ -48,7 +46,7 @@ class ActionResolver
     ): ActionInterface {
         if ($action instanceof InteractionActionInterface) {
             $identifier = $action->getIdentifier();
-            $resolvedIdentifier = $this->resolveValue($identifier, $pageProvider, $identifierProvider);
+            $resolvedIdentifier = $this->elementResolver->resolve($identifier, $pageProvider, $identifierProvider);
 
             if ($resolvedIdentifier !== $identifier) {
                 $action = $action->withIdentifier($resolvedIdentifier);
@@ -57,7 +55,7 @@ class ActionResolver
 
         if ($action instanceof InputActionInterface) {
             $value = $action->getValue();
-            $resolvedValue = $this->resolveValue($value, $pageProvider, $identifierProvider);
+            $resolvedValue = $this->elementResolver->resolve($value, $pageProvider, $identifierProvider);
 
             if ($resolvedValue !== $value) {
                 $action = $action->withValue($resolvedValue);
@@ -65,32 +63,5 @@ class ActionResolver
         }
 
         return $action;
-    }
-
-    /**
-     * @param string $value
-     * @param PageProviderInterface $pageProvider
-     * @param IdentifierProviderInterface $identifierProvider
-     *
-     * @return string
-     *
-     * @throws UnknownIdentifierException
-     * @throws UnknownPageElementException
-     * @throws UnknownPageException
-     */
-    private function resolveValue(
-        string $value,
-        PageProviderInterface $pageProvider,
-        IdentifierProviderInterface $identifierProvider
-    ): string {
-        if (ElementReference::is($value)) {
-            return $identifierProvider->findIdentifier((new ElementReference($value))->getElementName());
-        }
-
-        if (PageElementReference::is($value)) {
-            return $this->pageElementReferenceResolver->resolve($value, $pageProvider);
-        }
-
-        return $value;
     }
 }
